Add phpmd.xml and phpstan.neon to .gitattributes (#1)
This commit adds phpmd.xml and phpstan.neon to the .gitattributes file to be excluded during the export process. This change ensures that these configuration files, necessary for code analysis tools PHPMD and PHPStan respectively, don't get added to the repository when creating an archive or snapshot of the codebase.