Add PHPStan workflow (#1)
This commit introduces a new workflow for PHPStan static code analysis. The workflow is set to run on any push or pull request on all branches. PHPStan will operate on three operating systems - Ubuntu, Windows, and macOS, with PHP versions 8.2 and 8.3 used for the analysis. The analysis results are then made available for further examination.